Option to load dictionaries from a Reader
The original JaSpell could only accept a file path from which to load
dictionaries, which meant (a) you can't load a dictionary from inside a JAR
file and (b) the files were always loaded using the platform default encoding.
By allowing the caller to pass in a Reader rather than a filename, both of
these problems are removed as it becomes possible to read data from anywhere
(including a file, URL, Class.getResourceAsStream, ...) and control the
character encoding used.